Migration change

+++ BodaBodaServer/Migrations/20181203155108_Initial.cs
@@ -32,7 +32,7 @@ namespace BodaBodaServer.Migrations
                 {
                     LocationId = table.Column<long>(nullable: false)
                         .Annotation("MySQL:AutoIncrement", true),
-                    Longitude = table.Column<double>(nullable: false),
+                    Longitude = table.Column<long>(nullable: false),
                     Latitude = table.Column<long>(nullable: false),
                     LocationType = table.Column<string>(nullable: false),
                     UserId = table.Column<long>(nullable: false),
